Refactoring and documentation; title screen; font system; complete game.

+++ inc/gbadma.h
@@ -1,8 +1,6 @@
 #ifndef H_GBADMA
 #define H_GBADMA  1
 
-// TODO: docs
-
 #define REG_DMA_SRC ((volatile void*) 0x40000B0)
 #define REG_DMA_DST ((volatile void*) 0x40000B4)
 #define REG_DMA_CNT ((volatile uint32_t*) 0x40000B8)
@@ -17,6 +15,12 @@
 #define DMA_HALFWORD_CHUNK 0
 #define DMA_WORD_CHUNK 1
 
-void dma_zero_memory(void* dst, uint16_t len);
+/**
+ * Zero memory routine using Direct Memory Access (DMA) hardware. Similar
+ * to cpu_zero_memory but this function halts the CPU.
+ * @param dst The destination address to being zeroing.
+ * @param len The length in words.
+ */
+void dma_zero_memory(void* dst, uint32_t len);
 
 #endif
